Moved all public exceptions into new sub-module 'exceptions', and changed ParseError to be derived from Error
